requirements.txt
opencv-python
Pillow>=9.0.1
streamlit>=1.24.1